requirements.txt
# MCP SDK
git+https://github.com/modelcontextprotocol/python-sdk.git

# File system and pattern matching
pathspec>=0.11.0
gitignore-parser>=0.1.3

# Git operations
GitPython>=3.1.30

# Code parsing
tree-sitter>=0.20.1

# Caching
diskcache>=5.4.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
